pyproject.toml
[project]
name = "autozenith"
version = "0.1"
dependencies = [
"altgraph==0.17.4",
"azure-core==1.32.0",
"azure-identity==1.20.0",
"bcrypt==4.3.0",
"certifi==2025.1.31",
"cffi==1.17.1",
"charset-normalizer==3.4.1",
"cryptography==44.0.1",
"idna==3.10",
"isodate==0.7.2",
"msal==1.31.1",
"msal-extensions==1.2.0",
"msrest==0.7.1",
"oauthlib==3.2.2",
"packaging==24.2",
"portalocker==2.10.1",
"pycparser==2.22",
"pydo==0.9.2",
"pyinstaller==6.12.0",
"pyinstaller-hooks-contrib==2025.1",
"PyJWT==2.10.1",
"PyNaCl==1.5.0",
"PyYAML==6.0.2",
"requests==2.32.3",
"requests-oauthlib==2.0.0",
"setuptools==75.8.2",
"six==1.17.0",
"typing_extensions==4.12.2",
"urllib3==2.3.0"
]